+++ src/tool/java/tool/DataBuilder.java
@@ -1,4 +1,4 @@
-package dbbuilder;
+package tool;
 
 public interface DataBuilder {
     void run() throws Exception;
